feat: 코드 아이템 관리 기능 추가 및 개선
- `CodeItem` 엔티티 및 관련 리포지토리(`CodeItemRepository`, `CodeItemTreeRepository`)를 추가하여 코드 아이템 관리 기능을 구현.
- `CodeItemService` 및 `CodeItemTreeService` 클래스를 추가하여 비즈니스 로직을 통합하고, REST API를 위한 컨트롤러(`CodeItemRestController`, `CodeItemController`)를 구현.
- `MyBatisConfiguration`에서 `JdbcTemplate`을 설정하여 데이터베이스 접근을 개선.
- `build.gradle` 파일을 수정하여 `javadoc-extractor` 모듈을 포함하고, 관련 의존성을 추가.
- 다국어 지원을 위한 메시지 리소스 파일을 추가하고, 기존 메시지 파일을 업데이트하여 한국어 및 일본어 지원을 강화.
- 사용자 역할 및 계정 관련 DTO 및 서비스에서 `itemOrder` 필드를 `Integer`로 변경하여 데이터 일관성을 유지.
- 테스트 케이스를 추가하여 새로운 기능의 동작을 검증하고, 코드 품질을 향상시킴.

+++ simplix-example/src/main/java/dev/simplecore/simplix/demo/domain/common/user/entity/UserOrganization.java
@@ -58,7 +58,7 @@ public class UserOrganization extends AuditingBaseEntity<String> {
     @Column(name = "item_order", nullable = false)
     @Comment("Order")
     @I18nTitle({"ko=순서", "en=Order", "ja=順서"})
-    private BigDecimal itemOrder;
+    private Integer itemOrder;
 
         
     //----------------------------------
