+++ simplix-example/src/main/java/dev/simplecore/simplix/demo/domain/common/user/entity/UserPosition.java
@@ -44,7 +44,7 @@ public class UserPosition extends AuditingBaseEntity<String> {
     @Column(name = "item_order", nullable = false)
     @Comment("Order")
     @I18nTitle({"ko=순서", "en=Order", "ja=順서"})
-    private BigDecimal itemOrder;
+    private Integer itemOrder;
 
         
     //----------------------------------
